requirements.txt
torch>=1.9.0
transformers>=4.15.0
numpy>=1.20.0
Pillow>=8.0.0
requests>=2.25.0
tqdm>=4.60.0
matplotlib>=3.4.0
fake-useragent>=0.1.11
